requirements.txt
Authlib==1.0.1
bcrypt==3.2.0
blinker==1.4
CacheControl==0.12.11
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
cryptography==37.0.2
distlib==0.3.4
docopt==0.6.2
envoy-webhook-auth-decorator==0.1.1
filelock==3.6.0
Flask==2.1.1
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-Login==0.6.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
flatten-json==0.1.13
google-auth==2.6.6
google-auth-oauthlib==0.5.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
msgpack==1.0.3
oauth==1.0.1
oauthlib==3.2.0
pipreqs==0.4.11
platformdirs==2.5.1
protobuf==3.19.4
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
python-dotenv==0.20.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.35
urllib3==1.26.9
virtualenv==20.14.1
Werkzeug==2.1.1
WTForms==3.0.1
yarg==0.1.9
zipp==3.8.0
